+++ src/components/common/Button/style.js
@@ -1,0 +1,11 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const ButtonStyled = makeStyles(() => ({
+  root: {
+    backgroundColor: 'blue',
+    width: '100%',
+    borderRadius: 10,
+  },
+}));
+
+export default ButtonStyled;