+++ next.config.js
@@ -1,72 +1,18 @@
-const withBundleAnalyzer = require('@next/bundle-analyzer');
-const nextBuildId = require('next-build-id');
+const withBundleAnalyzer = require('@next/bundle-analyzer')({
+  enabled: process.env.NEXT_PUBLIC_APP_ANALYZE === true || process.env.ANALYZE === true,
+});
 const withCSS = require('@zeit/next-css');
-
+const nextBuildId = require('next-build-id');
+const withPlugins = require('next-compose-plugins');
 const isProd = process.env.NEXT_PUBLIC_APP_ENV === 'production';
-
-// const compose = (plugins) => ({
-//   webpack(config, options) {
-//     return plugins.reduce((config, plugin) => {
-//       if (plugin instanceof Array) {
-//         const [_plugin, ...args] = plugin;
-//         plugin = _plugin(...args);
-//       }
-//       if (plugin instanceof Function) {
-//         plugin = plugin();
-//       }
-//       if (plugin && plugin.webpack instanceof Function) {
-//         return plugin.webpack(config, options);
-//       }
-//       return config;
-//     }, config);
-//   },
-
-//   webpackDevMiddleware(config) {
-//     return plugins.reduce((config, plugin) => {
-//       if (plugin instanceof Array) {
-//         const [_plugin, ...args] = plugin;
-//         plugin = _plugin(...args);
-//       }
-//       if (plugin instanceof Function) {
-//         plugin = plugin();
-//       }
-//       if (plugin && plugin.webpackDevMiddleware instanceof Function) {
-//         return plugin.webpackDevMiddleware(config);
-//       }
-//       return config;
-//     }, config);
-//   },
-
-//   // Indicator from next for prerendered page status
-//   devIndicators: {
-//     autoPrerender: !isProd,
-//   },
-
-//   // Setting a custom build id
-//   generateBuildId: () => nextBuildId({ dir: __dirname, describe: true }),
-
-//   // Disabling x-powered-by in response headers
-//   poweredByHeader: false,
-// });
-
-// module.exports = compose([
-//   [
-//     // Bundle analyzer
-//     withBundleAnalyzer,
-//     {
-//       enabled: process.env.ANALYZE === 'true',
-//     },
-//   ],
-// ]);
-
-module.exports = withCSS({
+const nextConfig = {
   webpack: function (config) {
     config.module.rules.push({
       test: /\.(eot|woff|woff2|ttf|svg|png|jpg|gif)$/,
       use: {
         loader: 'url-loader',
         options: {
-          limit: 1000000,
+          limit: 100000,
           name: '[name].[ext]',
         },
       },
@@ -83,4 +29,6 @@ module.exports = withCSS({
 
   // Disabling x-powered-by in response headers
   poweredByHeader: false,
-});
+};
+
+module.exports = withPlugins([[withBundleAnalyzer], [withCSS]], nextConfig);