+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Removed Graphql configs
 - Removed Redux example in pages folder
 - Removed Nprogress plugin
+- Removed common Alert component
 
 ### Added
 
@@ -17,6 +18,8 @@
 - Adding integration local MUI theme
 - Adding integration Redux configs
 - Adding assets image (logo, manifest logo, favicon)
+- Adding common Button component
+- Adding auto open browser tab with library [open](https://github.com/sindresorhus/open)
 
 ## 1.0.1 - Feb 2021